requirements.txt
asgiref==3.4.1
Django==3.2.8
Faker==9.5.2
django-bootstrap-pagination==1.7.1
python-dateutil==2.8.2
django-with-asserts==0.0.1
pytz==2021.3
humanize==3.13.1
six==1.16.0
sqlparse==0.4.2
text-unidecode==1.3
django-widget-tweaks==1.4.9
